refactor: spesialKeys have been moved to a separate file
